refactor: break off full sync and sending messages
