+++ packages/loot-core/src/server/sync/full.js
@@ -16,71 +16,13 @@ import * as encoder from './encoder';
 import { getServer } from '../server-config';
 import { rebuildMerkleHash } from './repair';
 import { checkSyncingMode } from "./syncing-mode";
-import { applyMessages, deserializeValue } from "./sync-apply";
+import { deserializeValue } from "./sync-apply";
 import { receiveMessages } from "./sync-receive";
 import { getTablesFromMessages, getMessagesSince } from './utils';
 
 const { PostError, SyncError } = require('../errors');
 
 let FULL_SYNC_DELAY = 1000;
-
-async function _sendMessages(messages) {
-  try {
-    await applyMessages(messages);
-  } catch (e) {
-    if (e instanceof SyncError) {
-      if (e.reason === 'invalid-schema') {
-        // We know this message came from a local modification, and it
-        // couldn't apply, which doesn't make any sense. Must be a bug
-        // in the code. Send a specific error type for it for a custom
-        // message.
-        app.events.emit('sync', {
-          type: 'error',
-          subtype: 'apply-failure'
-        });
-      } else {
-        app.events.emit('sync', { type: 'error' });
-      }
-    }
-
-    throw e;
-  }
-
-  await scheduleFullSync();
-}
-
-let IS_BATCHING = false;
-let _BATCHED = [];
-export async function batchMessages(func) {
-  if (IS_BATCHING) {
-    await func();
-    return;
-  }
-
-  IS_BATCHING = true;
-  let batched = [];
-
-  try {
-    await func();
-    // TODO: if it fails, it shouldn't apply them?
-  } finally {
-    IS_BATCHING = false;
-    batched = _BATCHED;
-    _BATCHED = [];
-  }
-
-  if (batched.length > 0) {
-    await _sendMessages(batched);
-  }
-}
-
-export async function sendMessages(messages) {
-  if (IS_BATCHING) {
-    _BATCHED = _BATCHED.concat(messages);
-  } else {
-    return _sendMessages(messages);
-  }
-}
 
 export function clearFullSyncTimeout() {
   if (syncTimeout) {
