+++ packages/loot-core/src/server/sync/index.js
@@ -1,23 +1,23 @@
-export {
-  batchMessages,
-  sendMessages,
-  clearFullSyncTimeout,
-  scheduleFullSync,
-  initialFullSync,
-  fullSync
-} from "./sync"
+
 export { 
   serializeValue,
   deserializeValue,
   addSyncListener,
   applyMessages,
 } from "./sync-apply"
+export { batchMessages, sendMessages } from "./sync-send"
 export { receiveMessages, syncAndReceiveMessages } from "./sync-receive"
 export { 
   checkSyncingMode,
   setSyncingMode,
 } from "./syncing-mode"
 export { default as makeTestMessage } from './make-test-message';
+export {
+  clearFullSyncTimeout,
+  scheduleFullSync,
+  initialFullSync,
+  fullSync
+} from "./full"
 export { default as resetSync } from './reset';
 export { default as repairSync } from './repair';
 export { getMessagesSince } from "./utils"
